requirements.txt
Flask==1.1.2
flask_cors==3.0.10
Pillow==8.4.0
pytesseract==0.3.5
